+++ src/main/kotlin/io/github/durun/nitron/inout/model/ast/table/Structures.kt
@@ -1,7 +1,7 @@
 package io.github.durun.nitron.inout.model.ast.table
 
+import io.github.durun.nitron.core.AstSerializers
 import io.github.durun.nitron.core.MD5
-import io.github.durun.nitron.core.ast.type.AstSerializers
 import io.github.durun.nitron.core.ast.type.NodeTypePool
 import io.github.durun.nitron.core.toBlob
 import io.github.durun.nitron.core.toMD5
