+++ src/test/kotlin/io/github/durun/nitron/core/ast/node/AstTerminalNodeTest.kt
@@ -1,6 +1,6 @@
 package io.github.durun.nitron.core.ast.node
 
-import io.github.durun.nitron.core.ast.type.AstSerializers
+import io.github.durun.nitron.core.AstSerializers
 import io.github.durun.nitron.core.ast.type.NodeTypePool
 import io.github.durun.nitron.core.ast.type.TokenType
 import io.kotest.core.spec.style.FreeSpec
