+++ src/test/kotlin/io/github/durun/nitron/core/ast/node/BasicAstRuleNodeTest.kt
@@ -1,6 +1,6 @@
 package io.github.durun.nitron.core.ast.node
 
-import io.github.durun.nitron.core.ast.type.AstSerializers
+import io.github.durun.nitron.core.AstSerializers
 import io.github.durun.nitron.core.ast.type.RuleType
 import io.github.durun.nitron.core.ast.type.TokenType
 import io.github.durun.nitron.core.ast.type.createNodeTypePool
